+++ src/zrb/callback/callback.py
@@ -5,6 +5,7 @@ from zrb.callback.any_callback import AnyCallback
 from zrb.session.any_session import AnySession
 from zrb.task.any_task import AnyTask
 from zrb.util.attr import get_str_dict_attr
+from zrb.util.string.conversion import to_snake_case
 from zrb.xcom.xcom import Xcom
 
 
@@ -14,21 +15,15 @@ class Callback(AnyCallback):
         task: AnyTask,
         input_mapping: StrDictAttr,
         render_input_mapping: bool = True,
-        inflight_queue: str | None = None,
+        result_queue: str | None = None,
     ):
         self._task = task
         self._input_mapping = input_mapping
         self._render_input_mapping = render_input_mapping
-        self._inflight_queue = inflight_queue
+        self._result_queue = result_queue
 
     async def async_run(self, parent_session: AnySession, session: AnySession) -> Any:
-        parent_xcom = parent_session.shared_ctx.xcom
-        # publish to inflight queue
-        if self._inflight_queue is not None:
-            if self._inflight_queue not in parent_xcom:
-                parent_xcom[self._inflight_queue] = Xcom([])
-            parent_xcom[self._inflight_queue].push("")
-        # prepare input and run
+        # prepare input
         inputs = get_str_dict_attr(
             session.shared_ctx,
             self._input_mapping,
@@ -36,9 +31,16 @@ class Callback(AnyCallback):
         )
         for name, value in inputs.items():
             session.shared_ctx.input[name] = value
+            session.shared_ctx.input[to_snake_case(name)] = value
+        # run task and get result
         result = await self._task.async_run(session)
-        # consume from inflight queue
-        if self._inflight_queue is not None:
-            parent_xcom[self._inflight_queue].pop()
-        # return the result
+        self.publish_result(parent_session, result)
         return result
+
+    def publish_result(self, parent_session: AnySession, result: Any):
+        # publish to result queue
+        parent_xcom = parent_session.shared_ctx.xcom
+        if self._result_queue is not None:
+            if self._result_queue not in parent_xcom:
+                parent_xcom[self._result_queue] = Xcom([])
+            parent_xcom[self._result_queue].push(result)